+++ frontend/models/patient.ts
@@ -64,3 +64,13 @@ export const DEFAULT_PATIENT = {
     ],
     isPrioritized: false,
 };
+
+export interface StatusDataProps {
+    StatusId: number;
+    Temperature: number;
+    StatusTime: string;
+    IsReviewed: boolean;
+    Patient_PatientId: number;
+    Weight: number;
+    Symptoms: string;
+}
